requirements.txt
streamlit==1.32.0
supabase==2.4.0
postgrest==0.16.1
pandas==2.2.1
numpy==1.26.4
pydeck==0.9.0
plotly==5.20.0
python-dotenv==1.0.1
httpx==0.27.0
Pillow==10.3.0
